# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Yeungjin-Global",
        "description": "YJG API Documentation",
        "version": "0.1"
    },
    "paths": {
...
        "/api/user": {
            "get": {
                "tags": [
                    "유저"
                ],
                "summary": "유저 정보",
                "description": "현재 인증된 유저의 정보를 반환",
                "operationId": "2183352f835b442102aaa60ac0d76c29",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "ServerError"
                    }
                }
            },
            "post": {
                "tags": [
                    "학생"
                ],
                "summary": "회원가입(인증X)",
                "description": "회원가입",
                "operationId": "86f09d8ce9fe52b8e6b777dd707385ad",
                "requestBody": {
                    "description": "회원가입 정보",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "사용자 이름",
                                        "type": "string",
                                        "example": "관리자"
                                    },
                                    "student_id": {
                                        "description": "학번",
                                        "type": "string",
                                        "example": "1901234"
                                    },
                                    "phone_number": {
                                        "description": "전화번호",
                                        "type": "string",
                                        "example": "01012345678"
                                    },
                                    "email": {
                                        "description": "이메일",
                                        "type": "string",
                                        "example": "admin@gmail.com"
                                    },
                                    "password": {
                                        "description": "비밀번호",
                                        "type": "string",
                                        "example": "admin123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "422": {
                        "description": "ValidationException"
                    },
                    "500": {
                        "description": "ServerError"
                    }
                }
            },
            "delete": {
                "tags": [
                    "학생"
                ],
                "summary": "탈퇴",
                "description": "학생 유저의 회원 탈퇴 시 사용합니다.",
                "operationId": "c256e11ea1b42e32fea712214efd4c4d",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "ServerError"
                    }
                }
            },
            "patch": {
                "tags": [
                    "학생"
                ],
                "summary": "개인정보 수정",
                "description": "유저의 개인정보 수정 시 사용합니다.",
                "operationId": "401c87ec9c114bfb23532be2e7ab6228",
                "requestBody": {
                    "description": "수정할 유저의 정보",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "student_id": {
                                        "description": "정보를 수정할 유저의 아이디",
                                        "type": "string",
                                        "example": 1
                                    },
                                    "name": {
                                        "description": "변경할 이름",
                                        "type": "string",
                                        "example": "hyun"
                                    },
                                    "phone_number": {
                                        "description": "변경할 휴대폰 번호",
                                        "type": "string",
                                        "example": "01012345678"
                                    },
                                    "current_password": {
                                        "description": "이전 비밀번호",
                                        "type": "string",
                                        "example": "asdf321"
                                    },
                                    "new_password": {
                                        "description": "변경할 비밀번호",
                                        "type": "string",
                                        "example": "asdf123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "ModelNotFoundException"
                    },
                    "422": {
                        "description": "ValidationException"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/user/google-login": {
            "post": {
                "tags": [
                    "학생"
                ],
                "summary": "로그인",
                "description": "유저 google 로그인 시 사용합니다.",
                "operationId": "ef19ddde5df8c5453e5425bdac9b35df",
                "requestBody": {
                    "description": "회원가입 정보",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "이메일",
                                        "type": "string",
                                        "example": "admin@gmail.com"
                                    },
                                    "displayName": {
                                        "description": "이름",
                                        "type": "string",
                                        "example": "엄준식"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "422": {
                        "description": "ValidationException"
                    },
                    "500": {
                        "description": "ServerError"
                    }
                }
            }
        },
        "/api/user/login": {
            "post": {
                "tags": [
                    "학생"
                ],
                "summary": "로그인",
                "description": "일반 로그인 시 사용합니다.",
                "operationId": "b361a49b7041e04422d9e27f6edc43d0",
                "requestBody": {
                    "description": "회원가입 정보",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "이메일",
                                        "type": "string",
                                        "example": "admin@gmail.com"
                                    },
                                    "password": {
                                        "description": "비밀번호",
                                        "type": "string",
                                        "example": "admin123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "422": {
                        "description": "ValidationException"
                    },
                    "500": {
                        "description": "ServerError"
                    }
                }
            }
        },
        "/api/user/logout": {
            "post": {
                "tags": [
                    "학생"
                ],
                "summary": "로그아웃",
                "description": "유저 로그아웃",
                "operationId": "8b589eead1accdf2271a4ebc07a06168",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/user/approve": {
            "patch": {
                "tags": [
                    "학생"
                ],
                "summary": "회원가입 승인",
                "description": "학생 회원가입 승인",
                "operationId": "713a7fe5d63265247af733f4a4ee9117",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "ModelNotFoundException"
                    },
                    "500": {
                        "description": "ServerError"
                    }
                }
            }
        },
        "/api/user/verify-email/{id}": {
            "get": {
                "tags": [
                    "학생"
                ],
                "summary": "이메일 중복 확인",
                "description": "학생 이메일 중복 여부를 확인할 때 사용합니다.",
                "operationId": "8e576b925607f3969b0e9119e3b1c1be",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "중복을 확인할 학생의 이메일",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "422": {
                        "description": "ValidationException"
                    },
                    "500": {
                        "description": "ServerError"
                    }
                }
            }
        },
        "/api/user/reset-password": {
            "post": {
                "tags": [
                    "학생"
                ],
                "summary": "비밀번호 초기화",
                "description": "회원가입 시 입력한 이름, 이메일을 검증하고, 메일 전송 후 코드를 인증",
                "operationId": "fd28ffa8acd728bc29c4dbfa2a212841",
                "requestBody": {
                    "description": "name & email",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "이름",
                                        "type": "string",
                                        "example": "testname"
                                    },
                                    "email": {
                                        "description": "이메일",
                                        "type": "string",
                                        "example": "test@test.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "ModelNotFoundException"
                    },
                    "422": {
                        "description": "ValidationException"
                    },
                    "500": {
                        "description": "ServerError"
                    }
                }
            }
        },
        "/api/user/password": {
            "patch": {
                "tags": [
                    "학생"
                ],
                "summary": "비밀번호 재설정",
                "description": "유저의 비밀번호 재설정 시 사용합니다.",
                "operationId": "a99eeb982e343cc05a2eedefe7af6337",
                "requestBody": {
                    "description": "수정할 유저의 비밀번호",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "password": {
                                        "description": "변경할 비밀번호",
                                        "type": "string",
                                        "example": "asdf123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "ModelNotFoundException"
                    },
                    "422": {
                        "description": "ValidationException"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
...
    }
}